requirements.txt
twython==3.2.0
pillow==2.6.1
